+++ timeutil/timeutil_test.go
@@ -367,7 +367,6 @@ func (s *TimeUtilSuite) TestHelpers(c *C) {
 	d = time.Date(2012, 1, 1, 12, 0, 0, 0, time.Local)
 	c.Assert(Since(d, DAY), Not(Equals), 0)
 
-	d = time.Date(2023, 8, 1, 6, 30, 15, 0, time.Local)
 	c.Assert(DurationAs(time.Duration(79872232972344474), DAY), Equals, 924)
 }